sprated the files each functions its how part
