+++ src/index.html
@@ -44,7 +44,8 @@
         </div>
       </div>
     </section>
+    <script type="module" src="getData.js" defer></script>
     <script type="module" src="showData.js" defer></script>
-    <script type="module" src="script.js"></script>
+    <script type="module" src="getRecipes.js"></script>
   </body>
 </html>
